+++ ts/packages/nostr-pubsub/src/fips-invwant-tcp-driver.ts
@@ -18,6 +18,7 @@ import {
   type FipsInvWantTcpDriverOptions,
   type FipsInvWantTcpQueueSnapshot,
 } from './fips-invwant-tcp-types.js';
+import { abortTcpConnectionIfPresent } from './fips-tcp-cleanup.js';
 import { PubsubError, type NostrVerifiedEvent } from './types.js';
 
 const STREAM_IO_CHUNK_BYTES = 16 * 1024;
@@ -132,9 +133,10 @@ export class FipsInvWantTcpDriver {
       .filter(([, connection]) => connection.peer === peer)
       .map(([id]) => id);
     for (const id of ids) {
-      if (await this.tcp.state(id) !== undefined) {
-        await transport('abort TCP/FIPS pubsub peer', this.tcp.abort(id));
-      }
+      await transport(
+        'abort TCP/FIPS pubsub peer',
+        abortTcpConnectionIfPresent(this.tcp, id),
+      );
       this.connections.delete(id);
     }
     this.active.delete(peer);